Moved actions from tv-shows to media
