+++ app/actions/media.ts
@@ -16,6 +16,7 @@ import { getTvShowsApi } from "@jellyfin/sdk/lib/utils/api/tv-shows-api";
 import { getLibraryApi } from "@jellyfin/sdk/lib/utils/api/library-api";
 import { getGenresApi } from "@jellyfin/sdk/lib/utils/api/genres-api";
 import { createJellyfinInstance } from "@/lib/utils";
+import {UserItemDataDto} from "@jellyfin/sdk/lib/generated-client";
 
 // Type aliases for easier use
 type JellyfinItem = BaseItemDto;
@@ -180,6 +181,35 @@ export async function fetchMediaDetails(
 
     return null;
   }
+}
+
+export async function updatePlayedStatus(
+    item: JellyfinItem,
+    isPlayed: boolean,
+): Promise<UserItemDataDto | null> {
+    if (!item.Id) return null;
+
+    const { serverUrl, user } = await getAuthData();
+    const jellyfinInstance = createJellyfinInstance();
+    const api = jellyfinInstance.createApi(serverUrl);
+    api.accessToken = user.AccessToken;
+
+    try {
+        const itemsApi = getItemsApi(api);
+        const { data } = await itemsApi.updateItemUserData({
+            userId: user.Id,
+            itemId: item.Id,
+            updateUserItemDataDto: {
+                Played: isPlayed,
+                PlayedPercentage: isPlayed ? null : undefined,
+                PlaybackPositionTicks: isPlayed ? 0 : undefined
+            },
+        });
+        return data;
+    } catch (error) {
+        console.error("Failed to fetch item details:", error);
+        return null;
+    }
 }
 
 export async function fetchPersonDetails(
